+++ molecule/cli.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2015 Cisco Systems
+#  Copyright (c) 2015-2016 Cisco Systems
 #
 #  Permission is hereby granted, free of charge, to any person obtaining a copy
 #  of this software and associated documentation files (the "Software"), to deal
@@ -22,7 +22,7 @@ Usage:
     molecule [-hv] <command> [<args>...]
 
 Commands:
-    check         check playbook syntax
+    syntax        check playbook syntax
     create        create instances
     converge      create and provision instances
     idempotence   converge and check the output for changes
@@ -30,7 +30,7 @@ Commands:
     verify        create, provision and test instances
     destroy       destroy instances
     status        show status of instances
-    list          show available platforms, providers
+    list          show available platforms
     login         connects to instance via SSH
     init          creates the directory structure and files for a new Ansible role compatible with molecule
 
@@ -39,30 +39,29 @@ Options:
     -v --version  shows the version
 """
 
-import sys
-
-from docopt import docopt
-from docopt import DocoptExit
+import docopt
 
 import molecule
-import molecule.commands
+from molecule import commands
+from molecule import utilities
 
 
 class CLI(object):
     def main(self):
-        args = docopt(__doc__,
-                      version=molecule.__version__,
-                      options_first=True)
-        command_name = args.get('<command>').capitalize()
+        args = docopt.docopt(__doc__,
+                             version=molecule.__version__,
+                             options_first=True)
+        command_name = args.get('<command>')
         command_args = {} if args.get('<args>') is None else args.pop('<args>')
 
         try:
-            command_class = getattr(molecule.commands, command_name)
+            command_module = getattr(commands, command_name)
+            command_clazz = getattr(command_module, command_name.capitalize())
         except AttributeError:
-            raise DocoptExit()
+            raise docopt.DocoptExit()
 
-        c = command_class(command_args, args)
-        sys.exit(c.execute()[0])
+        c = command_clazz(command_args, args)
+        utilities.sysexit(c.execute()[0])
 
 
 def main():