+++ molecule/utilities.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2015 Cisco Systems
+#  Copyright (c) 2015-2016 Cisco Systems
 #
 #  Permission is hereby granted, free of charge, to any person obtaining a copy
 #  of this software and associated documentation files (the "Software"), to deal
@@ -20,14 +20,15 @@
 
 from __future__ import print_function
 
-import copy
+import logging
 import os
 import sys
-
-import logging
+import time
 
 import colorama
 import jinja2
+import m9dicts
+import paramiko
 
 
 class LogFilter(object):
@@ -45,66 +46,37 @@ class TrailingNewlineFormatter(logging.Formatter):
         return super(TrailingNewlineFormatter, self).format(record)
 
 
+colorama.init(autoreset=True)
+
 logger = logging.getLogger(__name__)
+
 warn = logging.StreamHandler()
 warn.setLevel(logging.WARN)
 warn.addFilter(LogFilter(logging.WARN))
-warn.setFormatter(TrailingNewlineFormatter('%(message)s'))
+warn.setFormatter(TrailingNewlineFormatter('{}%(message)s'.format(
+    colorama.Fore.YELLOW)))
 
 error = logging.StreamHandler()
 error.setLevel(logging.ERROR)
-error.setFormatter(TrailingNewlineFormatter('%(message)s'))
+error.setFormatter(TrailingNewlineFormatter('{}%(message)s'.format(
+    colorama.Fore.RED)))
 logger.addHandler(error)
 logger.addHandler(warn)
+logger.propagate = False
 
 
-def merge_dicts(a, b, raise_conflicts=False, path=None):
-    """
-    Merges the values of B into A.
-    If the raise_conflicts flag is set to True, a LookupError will be raised if the keys are conflicting.
-    :param a: the target dictionary
-    :param b: the dictionary to import
-    :param raise_conflicts: flag to raise an exception if two keys are colliding
-    :param path: the dictionary path. Used to show where the keys are conflicting when an exception is raised.
-    :return: The dictionary A with the values of the dictionary B merged into it.
-    """
-    # Set path.
-    if path is None:
-        path = []
+def print_success(msg):
+    print('{}{}'.format(colorama.Fore.GREEN, msg.rstrip()))
 
-    # Go through the keys of the 2 dictionaries.
-    for key in b:
-        # If the key exist in both dictionary, check whether we must update or not.
-        if key in a:
-            # Dig deeper for keys that have dictionary values.
-            if isinstance(a[key], dict) and isinstance(b[key], dict):
-                merge_dicts(a[key],
-                            b[key],
-                            raise_conflicts=raise_conflicts,
-                            path=(path + [str(key)]))
 
-            # Skip the identical values.
-            elif a[key] == b[key]:
-                pass
-            else:
-                # Otherwise raise an error if the same keys have different values.
-                if raise_conflicts:
-                    raise LookupError("Conflict at '{path}'".format(
-                        path='.'.join(path + [str(key)])))
-
-                # Or replace the value of A with the value of B.
-                a[key] = b[key]
-        else:
-            # If the key does not exist in A, import it.
-            a[key] = copy.deepcopy(b[key]) if isinstance(b[key],
-                                                         dict) else b[key]
-
-    return a
+def print_info(msg):
+    print('--> {}{}'.format(colorama.Fore.CYAN, msg.rstrip()))
 
 
 def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
     """
     Writes a file from a jinja2 template.
+
     :param src: the target template files to use
     :param dest: destination of the templatized file to be written
     :param kwargs: dictionary of arguments passed to jinja2 when rendering template
@@ -118,9 +90,8 @@ def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
 
     # template file doesn't exist
     if path and not os.path.isfile(src):
-        logger.error('\n{}Unable to locate template file: {}{}'.format(
-            colorama.Fore.RED, src, colorama.Fore.RESET))
-        sys.exit(1)
+        logger.error('\nUnable to locate template file: {}\n'.format(src))
+        sysexit()
 
     # look for template in filesystem, then molecule package
     loader = jinja2.ChoiceLoader([jinja2.FileSystemLoader(path,
@@ -137,6 +108,7 @@ def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
 def write_file(filename, content):
     """
     Writes a file with the given filename using the given content. Overwrites, does not append.
+
     :param filename: the target filename
     :param content: what gets written into the file
     :return: None
@@ -148,6 +120,7 @@ def write_file(filename, content):
 def format_instance_name(name, platform, instances):
     """
     Takes an instance name and formats it according to options specified in the instance's config.
+
     :param name: the name of the instance
     :param platform: the current molecule platform in use
     :param instances: the current molecule instances dict in use
@@ -171,7 +144,8 @@ def format_instance_name(name, platform, instances):
 
     # add platform to name
     if working_instance['options'].get('append_platform_to_hostname'):
-        return name + '-' + platform
+        if (platform != 'all'):
+            return name + '-' + platform
 
     # if we fall through, return the default name
     return name
@@ -208,9 +182,28 @@ def remove_args(command_args, args, kill):
     return new_command_args, new_args
 
 
+def reset_known_host_key(hostname):
+    return os.system('ssh-keygen -R {}'.format(hostname))
+
+
+def check_ssh_availability(hostip, user, timeout):
+    import socket
+    ssh = paramiko.SSHClient()
+    ssh.load_system_host_keys()
+    ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+    try:
+        ssh.connect(hostip, username=user)
+        return True
+    except (paramiko.BadHostKeyException, paramiko.AuthenticationException,
+            paramiko.SSHException, socket.error):
+        time.sleep(timeout)
+        return False
+
+
 def debug(title, data):
     """
     Prints colorized output for use when debugging portions of molecule.
+
     :param title: title of debug output
     :param data: data of debug output
     :return: None
@@ -220,3 +213,43 @@ def debug(title, data):
                    colorama.Back.RESET, colorama.Style.RESET_ALL]))
     print(''.join([colorama.Fore.BLACK, colorama.Style.BRIGHT, data,
                    colorama.Style.RESET_ALL, colorama.Fore.RESET]))
+
+
+def sysexit(code=1):
+    sys.exit(code)
+
+
+def merge_dicts(a, b):
+    """
+    Merges the values of B into A and returns a new dict.  Uses the same merge
+    strategy as ``config._combine``.
+
+    dict a
+    ------
+    b:
+       - c: 0
+       - c: 2
+    d:
+       e: "aaa"
+       f: 3
+
+    dict b
+    ------
+    a: 1
+    b:
+       - c: 3
+    d:
+       e: "bbb"
+
+    Will give an object such as:
+
+    {'a': 1, 'b': [{'c': 0}, {'c': 2}, {'c': 3}], 'd': {'e': "bbb", 'f': 3}}
+
+    :param a: the target dictionary
+    :param b: the dictionary to import
+    :return: dict
+    """
+    md = m9dicts.make(a, merge=m9dicts.MS_DICTS_AND_LISTS)
+    md.update(b)
+
+    return md