+++ tests/test_cli.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2015 Cisco Systems
+#  Copyright (c) 2015-2016 Cisco Systems
 #
 #  Permission is hereby granted, free of charge, to any person obtaining a copy
 #  of this software and associated documentation files (the "Software"), to deal
@@ -18,34 +18,23 @@
 #  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 #  THE SOFTWARE.
 
-import StringIO
+import docopt
+import pytest
 
-import testtools
-from testtools.matchers import MatchesRegex
-from docopt import DocoptExit
-from mock import patch
-
-from molecule.cli import CLI
+from molecule import cli
 
 
-class TestCLI(testtools.TestCase):
-    def setUp(self):
-        super(TestCLI, self).setUp()
-        self._cli = CLI()
+@pytest.fixture()
+def cli_instance():
+    return cli.CLI()
 
-    def test_cli_raises_usage_without_arguments(self):
-        self.assertRaises(DocoptExit, lambda: self._cli.main())
 
-    def test_cli_prints_version(self):
-        with patch('sys.argv', ['bin/molecule', '--version']):
-            with patch('sys.stdout', StringIO.StringIO()) as mocked_stdout:
-                result = self.assertRaises(SystemExit,
-                                           lambda: self._cli.main())
-                stdout = mocked_stdout.getvalue().strip()
+def test_cli_raises_usage_without_arguments(cli_instance):
+    with pytest.raises(docopt.DocoptExit):
+        cli_instance.main()
 
-                self.assertEqual(result.code, None)
-                self.assertThat(stdout, MatchesRegex('^\d\.\d\.\d'))
 
-    def test_cli_raises_usage_with_invalid_sub_command(self):
-        with patch('sys.argv', ['bin/molecule', 'invalid-subcommand']):
-            self.assertRaises(DocoptExit, lambda: self._cli.main())
+def test_cli_raises_usage_with_invalid_sub_command(cli_instance, mocker):
+    mocker.patch('sys.argv', ['bin/molecule', 'invalid-subcommand'])
+    with pytest.raises(docopt.DocoptExit):
+        cli_instance.main()